requirements.txt
python-telegram-bot==13.7
PyPDF2==2.12.1
reportlab==3.6.12